Revert "add config and change to multi platform"
This reverts commit db666027b3b7e25d479ed073eb9a1081d2609ad7.
